DONE:- solve buges

+++ src/controller/profile_controller.js
@@ -2,14 +2,23 @@ const ApiError = require("../utils/error")
 const EmployeeModel = require("../model/employee_model")
 const cloudinary = require("../utils/cloudinary");
 const fs = require("fs");
+const { FOLDER_NAME } = require("../config/string");
 
 async function addProfile(req, res, next) {
     try {
-        const path = req.file.path;
-        const result = await cloudinary.uploader.upload(path);
-        req.body.profilePic = result.secure_url;
-        req.body.publicId = result.public_id;
-        fs.unlinkSync(path);
+        delete req.body.email;
+        delete req.body.password;
+        const file = req.file;
+        if (file) {
+            const path = file.path;
+            const result = await cloudinary.uploader.upload(path, { folder: FOLDER_NAME });
+            req.body.profilePic = result.secure_url;
+            req.body.publicId = result.public_id;
+            if (req.user.publicId) {
+                await cloudinary.uploader.destroy(req.user.publicId);
+            }
+            fs.unlinkSync(path);
+        }
         const id = req.id
         const profile = await EmployeeModel.findByIdAndUpdate({ _id: id }, { $set: req.body }, { new: true })
         res.status(200).json({ statusCode: 200, success: true, data: profile, message: "profile details added successfully" })
@@ -18,13 +27,4 @@ async function addProfile(req, res, next) {
     }
 }
 
-async function getProfile(req, res, next) {
-    try {
-        const id = req.id
-        const profile = await EmployeeModel.findById({ _id: id })
-        res.status(200).json({ statusCode: 200, success: true, data: profile })
-    } catch (e) {
-        next(new ApiError(400, e.message))
-    }
-}
 module.exports = { addProfile }
